Starting to switch configs of RL algorithms to use more fluent builder patterns. Many parameter choices in different algorithms default to SOTA and only be changed in specific cases

+++ rl4j/rl4j-core/src/main/java/org/deeplearning4j/rl4j/learning/ILearning.java
@@ -42,7 +42,7 @@ public interface ILearning<O, A, AS extends ActionSpace<A>> {
 
     interface LConfiguration {
 
-        Integer getSeed();
+        Long getSeed();
 
         int getMaxEpochStep();
